refactor: move the activity class out of the flow package
The flow package in the app layer should no longer exist, as only the activity exists there. Since
the activity is hosting the application move it alongside the application class for better
visibility
