+++ app/src/androidTest/java/com/chesire/nekome/flow/series/SortTests.kt
@@ -2,8 +2,8 @@ package com.chesire.nekome.flow.series
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.rule.ActivityTestRule
+import com.chesire.nekome.Activity
 import com.chesire.nekome.R
-import com.chesire.nekome.flow.Activity
 import com.chesire.nekome.helpers.injector
 import com.chesire.nekome.helpers.login
 import com.chesire.nekome.kitsu.AuthProvider
