+++ app/src/main/AndroidManifest.xml
@@ -39,7 +39,7 @@
             android:value="598.0dip" />
 
         <activity
-            android:name="com.chesire.nekome.flow.Activity"
+            android:name="com.chesire.nekome.Activity"
             android:theme="@style/AppTheme.Launcher"
             android:windowSoftInputMode="adjustResize" >
             <intent-filter>
