+++ app/src/main/res/layout/activity.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.chesire.nekome.flow.Activity"
+    tools:context="com.chesire.nekome.Activity"
     tools:openDrawer="start">
 
     <include
