+++ app/src/main/res/layout/view_content.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.chesire.nekome.flow.Activity"
+    tools:context="com.chesire.nekome.Activity"
     tools:showIn="@layout/view_app_bar">
 
     <fragment
